Rename DocumentGeneratorConsumerApplicationConfiguration to EnvironmentConfiguration and add tests
