+++ src/main/java/uk/gov/companieshouse/document/generator/consumer/configuration/EnvironmentConfiguration.java
@@ -1,15 +1,21 @@
-package uk.gov.companieshouse.document.generator.consumer;
+package uk.gov.companieshouse.document.generator.consumer.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
 import uk.gov.companieshouse.environment.EnvironmentReader;
 import uk.gov.companieshouse.environment.impl.EnvironmentReaderImpl;
 
 @Configuration
-public class DocumentGeneratorConsumerApplicationConfiguration {
+public class EnvironmentConfiguration {
 
     @Bean
     EnvironmentReader environmentReader() {
         return new EnvironmentReaderImpl();
+    }
+
+    @Bean
+    RestTemplate restTemplate() {
+        return new RestTemplate();
     }
 }
