+++ src/main/java/uk/gov/companieshouse/document/generator/consumer/exception/GenerateDocumentException.java
